+++ scripts/compile.js
@@ -8,7 +8,9 @@ const pkg = argv.pkg || '*';
 
 const writeJs = require('./compile/write-js');
 const transformJs = require('./compile/transform-js');
+const sortFolders = require('./compile/sort-folders');
 const simpleCopy = require('./compile/simple-copy');
+const extractStyles = require('./compile/extract-styles');
 const { emitDeclarations, createTSProgram } = require('./compile/emit-declarations');
 
 (async function () {
@@ -16,23 +18,28 @@ const { emitDeclarations, createTSProgram } = require('./compile/emit-declaratio
   let tsTiming = 0;
   let emitTiming = 0;
   let simpleCopyTiming = 0;
+  let stylesTiming = 0;
 
   console.log(`Compiling...`);
 
   const packages = `../packages/${pkg}`;
 
   const folders = glob.sync(`${path.resolve(__dirname, packages)}`);
+  const linariaConfig = path.resolve(__dirname, '../linaria.config.js');
   const distPart = 'dist';
   const srcPart = 'src';
 
   const tsStart = moment();
   const tsFiles = glob.sync(path.resolve(__dirname, `../packages/*/src/**/*.{ts,tsx}`));
 
-  createTSProgram({ fileNames: tsFiles });
+  await createTSProgram({ fileNames: tsFiles });
 
   tsTiming += moment().diff(tsStart, 'ms');
 
-  for (const folder of folders) {
+  const sortedFolders = sortFolders(folders);
+
+  for (const folder of sortedFolders) {
+    const package = require(`${folder}/package.json`);
     const src = `${folder}/${srcPart}`;
     const dist = `${folder}/${distPart}`;
     const distESM = `${dist}/esm`;
@@ -48,7 +55,8 @@ const { emitDeclarations, createTSProgram } = require('./compile/emit-declaratio
 
     const copyStart = moment();
     const filesToCopy = glob.sync(`${src}/**/*.{woff,woff2,png}`);
-    filesToCopy.forEach(simpleCopy({ src, distCJS, distESM }));
+
+    filesToCopy.forEach(simpleCopy({ src, dist, distCJS, distESM }));
     simpleCopyTiming += moment().diff(copyStart, 'ms');
 
     const emitStart = moment();
@@ -61,12 +69,26 @@ const { emitDeclarations, createTSProgram } = require('./compile/emit-declaratio
     });
 
     emitTiming += moment().diff(emitStart, 'ms');
+
+    const stylesStart = moment();
+
+    extractStyles({
+      files: jsFiles,
+      configFile: linariaConfig,
+      version: package.version,
+      distESM,
+      distCJS,
+      src,
+    });
+
+    stylesTiming += moment().diff(stylesStart, 'ms');
   }
 
-  const totalTiming = tsTiming + jsTiming + emitTiming + simpleCopyTiming;
+  const totalTiming = tsTiming + jsTiming + emitTiming + simpleCopyTiming + stylesTiming;
   logInfo(`TS build has taken ${tsTiming}ms.`);
   logInfo(`JS build has taken ${jsTiming}ms.`);
   logInfo(`Copying files has taken ${simpleCopyTiming}ms.`);
   logInfo(`Types emitting has taken ${emitTiming}ms.`);
+  logInfo(`Styles extraction has taken ${stylesTiming}ms.`);
   logHelp(`Total ${totalTiming}ms.`);
 })();